Second commit

+++ admin/includes/view_all_posts.php
@@ -18,6 +18,28 @@
                    $query = "DELETE FROM comments WHERE comment_post_id = {$postIdValue} ";
                    $delete_comment_query = mysqli_query($conn,$query);
                   break;
+              case 'clone':
+                  $query = "SELECT * FROM posts WHERE post_id = {$postIdValue} ";
+                  $select_post_query = mysqli_query($conn,$query);
+                  while($row = mysqli_fetch_array($select_post_query)){
+                     $post_author = $row['post_author'];
+                     $post_title = $row['post_title'];
+                     $post_category_id = $row['post_category_id'];
+                     $post_status = $row['post_status'];
+                     $post_tags = $row['post_tags'];
+                     $post_image = $row['post_image'];
+                     $post_content = $row["post_content"];
+                     $post_comment_count = $row['post_comment_count'];
+                     $post_date = $row['post_date'];
+                      
+                      $query = "INSERT INTO posts(post_category_id,post_title,post_author,post_date,post_image,post_content,post_tags,post_comment_count,post_status) ";
+                      $query .= "VALUES({$post_category_id},'{$post_title}','{$post_author}',now(),'{$post_image}','{$post_content}','{$post_tags}','{$post_comment_count}','{$post_status}' )";
+
+                      $create_post_query = mysqli_query($conn,$query);
+
+                      confirmQuery($create_post_query);
+                  }
+                  
           }
       }
   }
@@ -35,6 +57,7 @@
                <option value="published">Publish</option>
                <option value="draft">Draft</option>
                <option value="delete">Delete</option>
+               <option value="clone">Clone</option>
            </select>
        </div>
        <div class="col-xs-4">
@@ -52,9 +75,10 @@
             <th>Tags</th>
             <th>Comments</th>
             <th>Date</th>
+            <th>Views</th>
         </tr>
         <?php
- $query = "select * from posts";
+ $query = "select * from posts order by post_id desc ";
  $select_all = mysqli_query($conn,$query);
 
  while($row = mysqli_fetch_assoc($select_all)){
@@ -67,6 +91,7 @@
      $post_image = $row['post_image'];
      $post_comment_count = $row['post_comment_count'];
      $post_date = $row['post_date'];
+     $post_views_count = $row['post_views_count'];
 
      echo "<tr>";
      ?>
@@ -87,8 +112,11 @@
      echo "<td> <img src='../images/$post_image' width='100' height='50' alt='image'></td>";
      echo "<td>$post_comment_count</td>";
      echo "<td>$post_date</td>";
+     //echo "<td><a href='comments.php?source=post_comments&p_id={$post_id}'>$post_views_count</a></td>";
+     echo "<td><a href='comments.php?source=post_comments&p_id={$post_id}'>$post_views_count</a></td>";
      echo "<td><a href='posts.php?source=edit_post&p_id={$post_id}'>Edit</a></td>";
-     echo "<td><a href='posts.php?delete={$post_id}'>Delete</a></td>";
+     echo "<td><a onClick=\"javascript: return confirm('Are you sure? You want to delete.'); \" href='posts.php?delete={$post_id}'>Delete</a></td>";
+    
      echo"</tr>";
